fix: add missing clusters, fix categories, and support half-star ratings
Fixture Data:
- Add 21 missing cluster definitions to clusters.yaml (IDs 74, 82, 85,
  86, 97, 113, 114, 129, 155, 158, 336, 1036, 1105-1107, 1295-1296,
  1872-1875)
- Fix category for On/Off Plug-in Unit (266) and Dimmable Plug-In Unit
  (267): lighting → appliances

Documentation:
- Fix rate limiter config reference: rate_limiter.yaml → framework.yaml
- Add app:scores:rebuild command documentation
- Add DeviceScoreService, DclApiService, DatabaseService to key components
- Update data model to distinguish Doctrine entities from raw SQL tables
- Expand entity field documentation for Cluster and DeviceType

Star Rating Enhancement:
- Change star rating from int (1-5) to float (0.5-5) for half-star support
- Update scoring thresholds for finer granularity
- Add half-star CSS styling with visual split display
- Cap non-compliant devices at 2.5 stars instead of 2

+++ src/Dto/DeviceScore.php
@@ -14,7 +14,7 @@ final class DeviceScore
      */
     public function __construct(
         public readonly float $overallScore,
-        public readonly int $starRating,
+        public readonly float $starRating,
         public readonly bool $isCompliant,
         public readonly array $scoresByType,
         public readonly ?string $bestVersion = null,
@@ -33,7 +33,7 @@ final class DeviceScore
 
         return new self(
             overallScore: (float) $data['overallScore'],
-            starRating: (int) $data['starRating'],
+            starRating: (float) $data['starRating'],
             isCompliant: (bool) $data['isCompliant'],
             scoresByType: $scoresByType,
             bestVersion: $data['bestVersion'] ?? null,
